backend/requirements/common.txt
django~=3.1
psycopg2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-pydenticon
pillow